+++ RELEASE_NOTES.md
@@ -11,6 +11,8 @@
 * GC-158 - Mailbox:processor list/read operation returns only a small subset of fields
 * GMB-1086 - Add Watchdog cleanup to lens event log
 * GMB-1101 - Processors' retryInterval-property should be supported
+* GGT-1581 - PX1 US UAT and PROD properties update
+* GMB-658 Mailbox updated with new 'tenancyKey' and invalid 'serviceInstance_pguid'.
 
 5.2.13
 ---------------